requirements.txt
argon2-cffi==20.1.0
attrs==23.1.0
certifi==2020.12.5
cffi==1.14.4
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
crispy-forms-gds==0.2.2
defusedxml==0.7.1
dj-database-url==0.5.0
django-allauth==0.47.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-cleanup==8.0.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.21.6
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.0.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
MarkupSafe==2.1.3
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==23.1
Pillow==8.1.0
pycparser==2.20
Pygments==2.7.4
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
requests==2.26.0
requests-oauthlib==1.3.0
rpds-py==0.8.4
simplejson==3.19.1
six==1.15.0
soupsieve==2.1
swagger-spec-validator==3.0.3
threadpoolctl==2.1.0
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.7

